API

NoLimitHoldem — Module

NoLimitHoldem

A no-limit hold-em simulator.

Terminology

game a single "game", where players are dealt hands, winner(s) are declared once.

state a point or process in the game, including PreFlop, Flop, Turn, River.

round the process of each player deciding which actions to take, until no further actions are taking.

NoLimitHoldem.move_button! — Function

move_button!(table::Table)

Move the button to the next player on the table.

NoLimitHoldem.play — Function

play(::Game)

Play a game. Note that this method expects no cards (players and table) to be dealt.