API

TexasHoldem — Module

TexasHoldem

A no-limit Texas Holdem simulator.

Terminology

game a single "game", where players are dealt hands, winner(s) are declared once.

stage a point or process in the game, including PreFlop, Flop, Turn, River.

round the process of each player deciding which actions to take, until no further actions are taking.

TexasHoldem.raise_to! — Function

raise_to!(game::Game, player::Player, amt)

Raise bet, for the round, to amount amt. Example:

# Flop
Player[1] raise to 10 (`amt = 10`, contribute 10)
Player[2] raise to 20 (`amt = 20`, contribute 20)
Player[3] raise to 40 (`amt = 40`, contribute 40)
Player[1] raise to 80 (`amt = 80`, contribute 80-10=70)
Player[2] call
Player[3] call
# Turn
Player[1] raise to 1 (`amt = 1`, contribute 1)
Player[2] raise to 2 (`amt = 2`, contribute 2)
Player[3] raise to 4 (`amt = 4`, contribute 4)
Player[1] raise to 8 (`amt = 8`, contribute 8-1=7)
Player[2] call
Player[3] call

TexasHoldem.move_buttons! — Function

move_buttons!(table::Table)

Move the dealer, small blind, big blind, and first-to-act buttons to the next set of players.

TexasHoldem.play! — Function

play!(game::Game)

Play a hand.

TexasHoldem.tournament! — Function

tournament!(game::Game)

Play until a single player remains!