API

TexasHoldem — Module

TexasHoldem

A no-limit Texas Holdem simulator.

Terminology

game a single "game", where players are dealt hands, winner(s) are declared once.

round a point or process in the game, including PreFlop, Flop, Turn, River.

round the process of each player deciding which actions to take, until no further actions are taking.

Game-level functions

TexasHoldem.move_buttons! — Function

move_buttons!(table::Table)

Move the dealer, small blind, big blind, and first-to-act buttons to the next set of players.

This is an internal method.

TexasHoldem.play! — Function

play!(game::Game)

Play a game.

TexasHoldem.tournament! — Function

tournament!(game::Game)

Play until a single player remains!

Player type

TexasHoldem.AbstractStrategy — Type

AbstractStrategy

An abstract strategy type, which users must subtype when passing to players.

TexasHoldem.Human — Type

Human

A human (terminal input) strategy. Use this strategy if you want to play "live" games, and not for simulating games.

TexasHoldem.Bot5050 — Type

Bot5050

Bot5050 is a stochastic strategy that chooses all of its actions based on a coin flip.

This bot is used for fuzzing in the test suite and for quick-start game configurations for users.

TexasHoldem.Player — Type

Player

Contains the poker player information:

strategy

cards

bank roll

pot investment

TODO: some of these fields should be removed, since they're only needed for flow control logic.

Player actions

TexasHoldem.Action — Type

Action

An action type, to be returned by player_option during each players turn to act. This is an internal type, but documented for user understanding.

Users are expected to return an action from player_option by calling one of Action's convenience methods:

Fold

Check

Call

Raise

AllIn

TexasHoldem.Check — Function

Check()

The check action, to be returned from player_option, when a player wants to check.

TexasHoldem.Fold — Function

Fold()

The fold action, to be returned from player_option, when a player wants to fold.

TexasHoldem.Call — Function

Call(amt::Int)
Call(table::Table, player::Player)

The call action, should be returned from player_option. when a player wants to call amount amt.

Use call_amount to query how much is needed to call.

TexasHoldem.call_amount — Function

call_amount(table::Table, player::Player)

Return the amount to call inside player_option.

TexasHoldem.Raise — Function

Raise(amt::Int)

The raise action, should be returned from player_option. when a player wants to raise to amount amt.

Use valid_raise_range to query the valid range that they are allowed to raise.

TexasHoldem.AllIn — Function

AllIn(amt::Int)
AllIn(table::Table, player::Player)

The call action, should be returned from player_option. when a player wants to raise to amount amt.

When AllIn is a player's only raise option, valid_raise_range will return a unit range of br:br, which is the player's appropriate (i.e., allowable) all-in bet.

For example, if players A and B have 150 and 100 bank rolls respectively, and player B bets 100, then player A's all-in raise amount would be 50 (on-top of the 100 call). This number is determined automatically.

So, users may simply use AllIn(table, player), which enters last(valid_raise_range(table, player)) as the amount.

TexasHoldem.valid_raise_range — Function

valid_raise_range(table::Table, player::Player)

A UnitRange{Int} of valid raises. Note that sometimes the range is n:n when all-in is the only available option.

See valid_raise_range_simple in the test suite for a more verbose but simpler implementation.

Player options

TexasHoldem.player_option — Function

player_option(game::Game, player::Player, option::AbstractPlayerOptions)

Returns a valid action (see Action), given the possible option. TexasHoldem calls player_option for each player on the table during each round. This function is entirely where the strategy logic resides.

CheckRaiseFold

CallRaiseFold

CallAllInFold

CallFold